Substitute init.environ.rc
CM/Lineage uses its own init.environ.rc. However it is hard to overwrite the original build file, because it also creates ramdisk folders and links.
Therefore include init.environ.rc directives are replaced by its code. This is not the best solution, but one we can live with.

We have to use the original init.environ.rc because it includes MTK jars the phone needs to include in order to work properly